Application now use prototype model
Backward incompatible changes
1) app.setupSlot deprecated. Use 'slotInit' event
2) app.registry now object. Use app.registry.hash method instead
3) More simplicity including application in end-point products: require('slot/app')
4) app instance don't exposed to global ns now
