+++ lib/stuff.js
@@ -7,7 +7,7 @@ var _ = require('lodash');
 /**
  * @private
  *
- * @param {string} cannel
+ * @param {string} channel
  * @returns {boolean}
  */
 function logEnabledFor(channel) {
@@ -46,8 +46,6 @@ function logToChannel(channel, type, args) {
     }
 }
 
-var stuff = module.exports = {
-
     /**
      * Форматирует строку.
      *
@@ -58,128 +56,130 @@ var stuff = module.exports = {
      * @param {...Array} values - Значения для форматирования.
      * @returns {string} Отформатированная строка.
      */
-    format: function(str) {
-        var values =  arguments;
+exports.format = function(str) {
+    var values =  arguments;
 
-        return str.replace(/%(\d+)/g, function(m, num) {
-            num = parseInt(num, 10);
-            return values[num];
-        });
-    },
+    return str.replace(/%(\d+)/g, function(m, num) {
+        num = parseInt(num, 10);
+        return values[num];
+    });
+};
 
-    /**
-     * Логирует в заданный канал.
-     *
-     * @param {string} channel - Канал.
-     * @param {...string} args - Аргументы сообщения.
-     */
-    log: function(channel, args) {
-        args = [].slice.call(arguments);
-        args.splice(1, 0, 'log');
-        logToChannel.apply(this, args);
-    },
+exports.escapeRegExp = function(str) {
+    return str.replace(/([.*+?^=!:${}()|\[\]\/\\])/g, "\\$1");
+};
 
-    /**
-     * @function
-     *
-     * @param {string} cannel
-     * @returns {boolean}
-     */
-    logEnabledFor: logEnabledFor,
+/**
+ * Логирует в заданный канал.
+ *
+ * @param {string} channel - Канал.
+ * @param {...string} args - Аргументы сообщения.
+ */
+exports.log = function(channel, args) {
+    args = [].slice.call(arguments);
+    args.splice(1, 0, 'log');
+    logToChannel.apply(this, args);
+};
 
-    /**
-     * Логирует в канал `warn`.
-     *
-     * @param {...string} args - Аргументы сообщения.
-     */
-    warn: function(args) {
-        args = [].slice.call(arguments);
-        args.unshift('warn', 'warn');
-        logToChannel.apply(this, args);
-    },
+/**
+ * @function
+ *
+ * @param {string} cannel
+ * @returns {boolean}
+ */
+exports.logEnabledFor = logEnabledFor;
 
-    /**
-     * Логирует в канал `error`.
-     *
-     * @param {...string} args - Аргументы сообщения.
-     */
-    error: function(args) {
-        args = [].slice.call(arguments);
-        args.unshift('error', 'error');
-        logToChannel.apply(this, args);
-    },
+/**
+ * Логирует в канал `warn`.
+ *
+ * @param {...string} args - Аргументы сообщения.
+ */
+exports.warn = function(args) {
+    args = [].slice.call(arguments);
+    args.unshift('warn', 'warn');
+    logToChannel.apply(this, args);
+};
 
-    /**
-     * Преобразует первый символ строки в заглавный.
-     *
-     * @param {string} str - Исходная строка.
-     * @returns {string}
-     */
-    capitalize: function(str) {
-        str = str == null ? '' : String(str);
-        return str.charAt(0).toUpperCase() + str.slice(1);
-    },
+/**
+ * Логирует в канал `error`.
+ *
+ * @param {...string} args - Аргументы сообщения.
+ */
+exports.error = function(args) {
+    args = [].slice.call(arguments);
+    args.unshift('error', 'error');
+    logToChannel.apply(this, args);
+};
 
-    /**
-     * Удаляет с конца массива переданное значение.
-     *
-     * @param {Array} array
-     * @param {*} value
-     */
-    pop: function(array, value) {
-        var lastIndex = _.lastIndexOf(array, value);
+/**
+ * Преобразует первый символ строки в заглавный.
+ *
+ * @param {string} str - Исходная строка.
+ * @returns {string}
+ */
+exports.capitalize = function(str) {
+    str = str == null ? '' : String(str);
+    return str.charAt(0).toUpperCase() + str.slice(1);
+};
 
-        if (lastIndex != -1) {
-            array.splice(lastIndex, 1);
-        }
-    },
+/**
+ * Удаляет с конца массива переданное значение.
+ *
+ * @param {Array} array
+ * @param {*} value
+ */
+exports.pop = function(array, value) {
+    var lastIndex = _.lastIndexOf(array, value);
 
-
-    /**
-     * Get property for object by path like `prop1.prop2.prop3`.
-     *
-     * @param {Object} obj
-     * @param {string} path
-     * @param {string} [separator=.] - Separotor for splitting path.
-     * @returns {*}
-     */
-    getByPath: function(obj, path, separator) {
-        separator = separator || '.';
-        var paths = path.split(separator);
-
-        while (paths.length && obj) {
-            obj = obj[paths.shift()];
-        }
-
-        return obj;
-    },
-
-    state2uri: function(state, encode) {
-        var parts = [];
-        for (var key in state) {
-            var value = state[key];
-            if (encode || encode == null) {
-                value = encodeURIComponent(value);
-            }
-            parts.push(key + '=' + value);
-        }
-        return parts.join('&');
-    },
-
-    uri2state: function(uri, decode) {
-        if (uri.charAt(0) == '?') uri = uri.substr(1);
-        var parts = uri.split('&');
-        var result = {};
-        _.each(parts, function(part) {
-            if (!part) return;
-            part.replace(/([^=]+)\=?(.*)/g, function(m, name, value) {
-                if (decode || decode == null) {
-                    value = decodeURIComponent(value);
-                }
-                result[name] = value;
-            });
-        });
-
-        return result;
+    if (lastIndex != -1) {
+        array.splice(lastIndex, 1);
     }
+};
+
+/**
+ * Get property for object by path like `prop1.prop2.prop3`.
+ *
+ * @param {Object} obj
+ * @param {string} path
+ * @param {string} [separator=.] - Separotor for splitting path.
+ * @returns {*}
+ */
+exports.getByPath = function(obj, path, separator) {
+    separator = separator || '.';
+    var paths = path.split(separator);
+
+    while (paths.length && obj) {
+        obj = obj[paths.shift()];
+    }
+
+    return obj;
+};
+
+exports.state2uri = function(state, encode) {
+    var parts = [];
+    for (var key in state) {
+        var value = state[key];
+        if (encode || encode == null) {
+            value = encodeURIComponent(value);
+        }
+        parts.push(key + '=' + value);
+    }
+    return parts.join('&');
+};
+
+exports.uri2state = function(uri, decode) {
+    if (uri.charAt(0) == '?') uri = uri.substr(1);
+    var parts = uri.split('&');
+    var result = {};
+    _.each(parts, function(part) {
+        if (!part) return;
+        part.replace(/([^=]+)\=?(.*)/g, function(m, name, value) {
+            if (decode || decode == null) {
+                value = decodeURIComponent(value);
+            }
+            result[name] = value;
+        });
+    });
+
+    return result;
 };
