+++ package.json
@@ -6,7 +6,8 @@
   "description": "Flexible framework for complex JS applications",
   "main": "index.js",
   "dependencies": {
-    "async": "0.2.9",
+    "async": "^0.9.0",
+    "async.emitter": "^1.0.0",
     "chokidar": "~0.8.2",
     "event-stream": "~3.0.16",
     "flat-glob": "0.0.1",
@@ -18,13 +19,14 @@
     "gulp-remember": "~0.2.0",
     "gulp-util": "~2.2.16",
     "handlebars": "~1.3.0",
+    "inherits": "^2.0.1",
     "jquery-deferred": "~0.3.0",
     "lodash": "2.4.1",
     "mkdirp": "~0.5.0",
     "po2json": "0.2.3",
+    "posix": "~1.0.3",
     "request": "~2.27.0",
     "setimmediate": "^1.0.2",
-    "smokesignals": "file:smokesignals",
     "ua-parser-js": "~0.6.2",
     "winston": "~0.7.2"
   },
@@ -57,6 +59,7 @@
   "browser": {
     "jquery-deferred": false,
     "request": false,
-    "./logger": false
+    "./logger": false,
+    "./app/index.js": "./app/clientApp.js"
   }
 }
