--- serverApp.js
@@ -1,14 +1,0 @@
-
-var baseAppConstructor = require('./app');
-
-/**
- * @class slot.ServerApp
- * @extends {slot.App}
- */
-module.exports = function(params) {
-    var baseApp = baseAppConstructor({
-        rootPath: params.rootPath + '/'
-    });
-
-    return baseApp.instance;
-};
