--- smokesignals/index.js
@@ -1,80 +1,0 @@
-// in a few cases we've chosen optimizing script length over efficiency of code.
-// I think that is the right choice for this library.  If you're adding and
-// triggering A LOT of events, you might want to use a different library.
-module.exports = {
-    convert: function(obj, handlers) {
-        // we store the list of handlers as a local variable inside the scope
-        // so that we don't have to add random properties to the object we are
-        // converting. (prefixing variables in the object with an underscore or
-        // two is an ugly solution)
-        //      we declare the variable in the function definition to use two less
-        //      characters (as opposed to using 'var ').  I consider this an inelegant
-        //      solution since smokesignals.convert.length now returns 2 when it is
-        //      really 1, but doing this doesn't otherwise change the functionallity of
-        //      this module, so we'll go with it for now
-        handlers = {};
-
-        // add a listener
-        obj.on = function(eventName, handler) {
-            // either use the existing array or create a new one for this event
-            //      this isn't the most efficient way to do this, but is the shorter
-            //      than other more efficient ways, so we'll go with it for now.
-            (handlers[eventName] = handlers[eventName] || [])
-                // add the handler to the array
-                .push(handler);
-
-            return obj;
-        };
-
-        // add a listener that will only be called once
-        obj.once = function(eventName, handler) {
-            // create a wrapper listener, that will remove itself after it is called
-            function wrappedHandler() {
-                // remove ourself, and then call the real handler with the args
-                // passed to this wrapper
-                handler.apply(obj.off(eventName, wrappedHandler), arguments);
-            }
-            // in order to allow that these wrapped handlers can be removed by
-            // removing the original function, we save a reference to the original
-            // function
-            wrappedHandler.h = handler;
-
-            // call the regular add listener function with our new wrapper
-            return obj.on(eventName, wrappedHandler);
-        };
-
-        // remove a listener
-        obj.off = function(eventName, handler) {
-            // loop through all handlers for this eventName, assuming a handler
-            // was passed in, to see if the handler passed in was any of them so
-            // we can remove it
-            //      it would be more efficient to stash the length and compare i
-            //      to that, but that is longer so we'll go with this.
-            for (var list = handlers[eventName], i = 0; handler && list && list[i]; i++) {
-                // either this item is the handler passed in, or this item is a
-                // wrapper for the handler passed in.  See the 'once' function
-                list[i] != handler && list[i].h != handler || list.splice(i--, 1);
-            }
-            // if i is 0 (i.e. falsy), then there are no items in the array for this
-            // event name (or the array doesn't exist)
-            if (!i) {
-                // remove the array for this eventname (if it doesn't exist then
-                // this isn't really hurting anything)
-                delete handlers[eventName];
-            }
-            return obj;
-        };
-
-        obj.emit = function(eventName) {
-            // loop through all handlers for this event name and call them all
-            //      it would be more efficient to stash the length and compare i
-            //      to that, but that is longer so we'll go with this.
-            for (var list = handlers[eventName], i = 0; list && list[i];) {
-                list[i++].apply(obj, list.slice.call(arguments, 1));
-            }
-            return obj;
-        };
-
-        return obj;
-    }
-};
