--- smokesignals/package.json
@@ -1,9 +1,0 @@
-{
-  "name": "smokesignals",
-  "version": "1.0.0",
-  "description": "temporary fixed smokesignals",
-  "main": "index.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  }
-}
