--- tests/clientApp.spec.js
@@ -1,6 +1,0 @@
-describe('app', function() {
-    it('Подключается без ошибок', function(done) {
-        require('../clientApp/index.js');
-        done();
-    });
-});
